+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Repository/PacienteRepository.java
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {
     List<Paciente> findByDni(String dni);
     List<Paciente> findByNombreContainingIgnoreCase(String nombre);
     List<Paciente> findByNroAfiliado(String nroAfiliado);
-    List<Paciente> findByPlan_IdPlan(Integer idPlan);
-    List<Paciente> findByPlan_ObraSocial_IdObraSocial(Integer idObraSocial);
+    List<Paciente> findByPlan_IdPlan(Long idPlan);
+    List<Paciente> findByPlan_ObraSocial_IdObraSocial(Long idObraSocial);
     boolean existsByDni(String dni);
 }
